+++ src/Contracts/Requests/CreateCustomer.php
@@ -1,11 +1,11 @@
 <?php
 /**
- * Contains the UpdateClient interface.
+ * Contains the CreateCustomer interface.
  *
  * @copyright   Copyright (c) 2017 Attila Fulop
  * @author      Attila Fulop
  * @license     MIT
- * @since       2017-10-04
+ * @since       2017-10-02
  *
  */
 
@@ -13,7 +13,7 @@
 namespace Konekt\AppShell\Contracts\Requests;
 
 
-interface UpdateClient extends BaseRequest
+interface CreateCustomer extends BaseRequest
 {
 
 }