+++ src/Contracts/Requests/UpdateCustomer.php
@@ -1,11 +1,11 @@
 <?php
 /**
- * Contains the CreateClient interface.
+ * Contains the UpdateCustomer interface.
  *
  * @copyright   Copyright (c) 2017 Attila Fulop
  * @author      Attila Fulop
  * @license     MIT
- * @since       2017-10-02
+ * @since       2017-10-04
  *
  */
 
@@ -13,7 +13,7 @@
 namespace Konekt\AppShell\Contracts\Requests;
 
 
-interface CreateClient extends BaseRequest
+interface UpdateCustomer extends BaseRequest
 {
 
 }