+++ src/resources/views/customer/create.blade.php
@@ -1,7 +1,7 @@
 @extends('appshell::layouts.default')
 
 @section('title')
-    {{ __('Create new client') }}
+    {{ __('Create new customer') }}
 @stop
 
 @section('content')
@@ -11,18 +11,18 @@
         <div class="col-xl-9">
             <div class="card card-accent-success">
                 <div class="card-header">
-                    {{ __('Enter Client Details') }}
+                    {{ __('Enter Customer Details') }}
                 </div>
                 <div class="card-block">
 
-                    {!! Form::model($client, ['route' => 'appshell.client.store', 'autocomplete' => 'off']) !!}
+                    {!! Form::model($customer, ['route' => 'appshell.customer.store', 'autocomplete' => 'off']) !!}
 
-                    @include('appshell::client._form')
+                    @include('appshell::customer._form')
 
                     <hr>
                     <div class="form-group">
-                        <button class="btn btn-success">{{ __('Create client') }}</button>
-                        <a href="{{ route('appshell.client.index') }}" class="btn btn-link text-muted">{{ __('Cancel') }}</a>
+                        <button class="btn btn-success">{{ __('Create customer') }}</button>
+                        <a href="{{ route('appshell.customer.index') }}" class="btn btn-link text-muted">{{ __('Cancel') }}</a>
                     </div>
 
                     {!! Form::close() !!}