+++ Changelog.md
@@ -2,6 +2,11 @@
 
 ## 0.9
 
+### Unreleased
+##### 2017-11-XX
+
+- Konekt Client library has been replaced with konekt/customer
+
 ### 0.9.0
 ##### 2017-11-24